Consolidate LocalGroup and LocalCacheGroup, since these are now identical.
